+++ core/agent_harness/turns/evidence_driver.py
@@ -174,7 +174,7 @@ def _load_gather_llm_or_none(error_reporter: ErrorReporter | None) -> Any | None
     client isn't available (unsupported provider, misconfig), the caller
     surfaces a controlled fallback rather than a hard error.
     """
-    from core.llm.agent_llm_client import get_agent_llm
+    from core.llm.factory import LLMRole, get_llm
 
     return _safe_execute(
         get_agent_llm,
