+++ tests/core/runtime/llm/test_llm_client_provider_registry.py
@@ -55,7 +55,7 @@ def test_create_llm_client_dispatches_registry_provider(
         **{f"{provider}_toolcall_model": "stub-model"},
     )
     monkeypatch.setattr("config.config.resolve_llm_settings", lambda: settings)
-    monkeypatch.setattr(llm_client, "resolve_llm_api_key", lambda _env_var: "test-key")
+    monkeypatch.setattr("core.llm.providers.provider_credentials.resolve_llm_api_key", lambda _env_var: "test-key")
 
     client = _create_llm_client("toolcall")
 
@@ -80,7 +80,7 @@ def test_create_llm_client_dispatches_registry_provider_to_litellm_when_transpor
         **{f"{provider}_toolcall_model": "stub-model"},
     )
     monkeypatch.setattr("config.config.resolve_llm_settings", lambda: settings)
-    monkeypatch.setattr(llm_client, "resolve_llm_api_key", lambda _env_var: "test-key")
+    monkeypatch.setattr("core.llm.providers.provider_credentials.resolve_llm_api_key", lambda _env_var: "test-key")
     monkeypatch.setenv("OPENSRE_LLM_TRANSPORT", "litellm")
 
     client = _create_llm_client("toolcall")
@@ -100,7 +100,7 @@ def test_create_llm_client_uses_sdk_without_transport_flag(
 ) -> None:
     settings = SimpleNamespace(provider="deepseek", deepseek_toolcall_model="deepseek-v4-flash")
     monkeypatch.setattr("config.config.resolve_llm_settings", lambda: settings)
-    monkeypatch.setattr(llm_client, "resolve_llm_api_key", lambda _env_var: "test-key")
+    monkeypatch.setattr("core.llm.providers.provider_credentials.resolve_llm_api_key", lambda _env_var: "test-key")
     monkeypatch.delenv("OPENSRE_LLM_TRANSPORT", raising=False)
 
     client = _create_llm_client("toolcall")
