+++ tools/investigation/stages/diagnose/node.py
@@ -97,7 +97,7 @@ def _parse_via_structured_output(
     *,
     alert_source: str = "",
 ) -> InvestigationResult:
-    from core.llm.llm_client import get_llm_for_reasoning
+    from core.llm.factory import LLMRole, get_llm
 
     prompt = f"""Extract the structured diagnosis from this investigation conclusion.
 
@@ -116,7 +116,7 @@ Evidence keys collected: {", ".join(evidence.keys()) if evidence else "none"}
         remediation_steps: list[str]
         validity_score: float
 
-    llm = get_llm_for_reasoning()
+    llm = get_llm(LLMRole.REASONING)
     schema_model = build_diagnosis_schema(taxonomy_categories_for_alert_source(alert_source))
     raw_schema = (
         llm.with_structured_output(schema_model)
