+++ tools/investigation/stages/intake/node.py
@@ -15,7 +15,7 @@ from core.domain.alerts.extraction import (
     make_problem_md,
 )
 from core.domain.types.incident_window import resolve_incident_window
-from core.llm.llm_client import get_llm_for_reasoning
+from core.llm.factory import LLMRole, get_llm
 from core.state import InvestigationState
 from platform.observability import (
     debug_print,
@@ -161,7 +161,7 @@ def _extract_alert_details(state: InvestigationState) -> AlertDetails:
     text = format_raw_alert(raw_alert)
     prompt = _EXTRACT_PROMPT.format(text=text)
 
-    llm = get_llm_for_reasoning()
+    llm = get_llm(LLMRole.REASONING)
     try:
         details = cast(
             AlertDetails,
